cout: rename sources
* Renamed cout.{hpp,cpp} -> lib/ostream.{hpp,cpp}

Activity: #39
