+++ include/lib/ostream.hpp
