+++ kernel/main.cpp
@@ -3,7 +3,6 @@
 #include <bolgenos-ng/error.h>
 #include <bolgenos-ng/time.h>
 
-#include <bolgenos-ng/cout.hpp>
 #include <bolgenos-ng/irq.hpp>
 #include <bolgenos-ng/mem_utils.hpp>
 #include <bolgenos-ng/memory.hpp>
@@ -16,6 +15,8 @@
 #include <bolgenos-ng/ps2.hpp>
 #include <bolgenos-ng/slab.hpp>
 #include <bolgenos-ng/vga_console.hpp>
+
+#include <lib/ostream.hpp>
 
 #include "config.h"
