+++ kernel/slab.cpp
@@ -2,9 +2,10 @@
 
 #include <bolgenos-ng/error.h>
 
-#include <bolgenos-ng/cout.hpp>
 #include <bolgenos-ng/mem_utils.hpp>
 #include <bolgenos-ng/memory.hpp>
+
+#include <lib/ostream.hpp>
 
 #include "config.h"
